wip: handle guard duty events

+++ packages/app-file-manager/src/modules/FileManagerRenderer/FileManagerViewProvider/FileManagerViewContext.tsx
@@ -17,10 +17,18 @@ import { ROOT_FOLDER } from "~/constants";
 
 type PublicState = Omit<State, "activeTags">;
 
+export interface UpdateFileOptions {
+    localUpdate?: boolean;
+}
+
+export interface DeleteFileOptions {
+    localDelete?: boolean;
+}
+
 export interface FileManagerViewContext<TFileItem extends FileItem = FileItem> extends PublicState {
     accept: string[];
     createFile: (data: TFileItem) => Promise<TFileItem | undefined>;
-    deleteFile: (id: string) => Promise<void>;
+    deleteFile: (id: string, options?: DeleteFileOptions) => Promise<void>;
     files: FileItem[];
     folderId: string;
     folders: FolderItem[];
@@ -59,7 +67,11 @@ export interface FileManagerViewContext<TFileItem extends FileItem = FileItem> e
     };
     toggleSelected: (file: TFileItem) => void;
     deselectAll: () => void;
-    updateFile: (id: string, data: Partial<TFileItem>) => Promise<void>;
+    updateFile: (
+        id: string,
+        data: Partial<TFileItem>,
+        options?: UpdateFileOptions
+    ) => Promise<void>;
     uploadFile: (file: File, options?: UploadFileOptions) => Promise<TFileItem | undefined>;
 }
 
@@ -255,11 +267,18 @@ export const FileManagerViewProvider = ({ children, ...props }: FileManagerViewP
         return newFile;
     };
 
-    const updateFile = async (id: string, data: Partial<FileItem>) => {
+    const updateFile = async (
+        id: string,
+        data: Partial<FileItem>,
+        options: UpdateFileOptions = {}
+    ) => {
         const file = files.find(file => file.id === id);
+
         if (!file) {
             return;
         }
+
+        const { localUpdate = false } = options;
 
         const newTags = data.tags?.filter(tag => !tags.tags.some(t => t.tag === tag));
 
@@ -271,7 +290,9 @@ export const FileManagerViewProvider = ({ children, ...props }: FileManagerViewP
             data.tags = [...(data.tags || []), props.scope];
         }
 
-        await fileManager.updateFile(id, data);
+        if (!localUpdate) {
+            await fileManager.updateFile(id, data);
+        }
 
         if (newTags) {
             tags.addTags(newTags);
@@ -279,6 +300,7 @@ export const FileManagerViewProvider = ({ children, ...props }: FileManagerViewP
 
         setFiles(files => {
             const index = files.findIndex(file => file.id === id);
+
             if (index === -1) {
                 return files;
             }
@@ -296,8 +318,12 @@ export const FileManagerViewProvider = ({ children, ...props }: FileManagerViewP
         });
     };
 
-    const deleteFile = async (id: string) => {
-        await fileManager.deleteFile(id);
+    const deleteFile = async (id: string, options: DeleteFileOptions = {}) => {
+        const { localDelete = false } = options;
+
+        if (!localDelete) {
+            await fileManager.deleteFile(id);
+        }
 
         setFiles(files => {
             const index = files.findIndex(file => file.id === id);
